pyproject.toml
[project]
name = "SegConv_CitySense"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "matplotlib>=3.9.4",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "scikit-learn>=1.6.1",
    "timm>=1.0.20",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]

[tool.uv.workspace]
members = [
    "segformer_task",
    "segformer",
]
